Start multiple teuthology workers
Allow multiple jobs to run in parallel by starting multiple teuthology
workers on the head node. Start n/3 workers, where n is the number of
test nodes in the system, on the assumption that a typical job requires
three worker nodes. Jobs that can't lock enough nodes will block until
they can, so it's not a problem to run too many workers, but it seems
wise to minimize contention anyway.