Clean up ansible plays per ansible-lint suggestions
In the head-node and test-node playbooks, it would be nice to change the
various ssh commands that modify stuff on the paddles node to use
`delegate_to`, letting ansible do the ssh work for us. But, when these
plays run ansible doesn't have the paddles node's host key. It would be
gross to disable host key checking entirely, so running ssh commands on
the nodes to update paddles is about all we can do.